+++ lib/td_audit_web/controllers/event_controller.ex
@@ -5,7 +5,7 @@ defmodule TdAuditWeb.EventController do
   alias TdAudit.Audit
   alias TdAuditWeb.SwaggerDefinitions
 
-  @filters_availables ["resource_id", "resource_type"]
+  @filters_availables ["resource_id", "resource_type", "event", "ts"]
 
   action_fallback(TdAuditWeb.FallbackController)